requirements.txt
python-dotenv==0.19.0
aiohttp==3.8.1
numpy==1.22.2
matplotlib==3.5.1
requests==2.26.0